Fixed gcc warnings when compiling console code with -Wall
Most of the fixes I applied here are pretty straightforward. Reordering member variable initialisation, making sure that we don't compare signed and unsigned ints, getting rid of two unused variables, that sort of thing.

There is one line that may deserve more scrutiny: the old legalchars[c]==arg.text line. I'm pretty sure that casting arg.text to a char is an appropriate way to resolve the signed vs. unsigned warning, and I'm pretty sure that a static cast is appropriate, but that one wasn't a COMPLETE no-brainer (i.e., if this were Wikipedia, I'd hesitate before marking that as a minor edit).