Texture Loading Utilizes Group Name
Append the group name provided to the TextureAtlas::_load function
to the "file" setting when the section is "texture".

Motivation:
The _loadTexture function is called without the group name
ever being considered. This results in the "General" group always being
selected, which in turn results in the Ogre::ResourceGroupManager doing
a case sensitive indexing, a case insensitive indexing, and a linear
search through the resource locations in the group.